+++ mobile/android/android-components/docs/changelog.md
@@ -16,6 +16,9 @@ permalink: /changelog/
   * ⚠️Increased `targetSdkVersion` to 30 (Android R)
 
 * **browser-toolbar**
+  * 🌟 Added `BrowserToolbarBehavior#forceCollapse` to easily collapse the top/bottom toolbar.
+
+* **browser-toolbar**
   * ⚠️ **This is a breaking change**: `BrowserToolbarBottomBehavior` is renamed to `BrowserToolbarBehavior` as it is now a common behavior for toolbars be them placed at the bottom or at the top of the screen.
 
 * **feature-session**